Fixed counting error with order

+++ src/Builder/Order.h
@@ -12,16 +12,16 @@ class Order {
         std::vector<OrderItem> orderItems;
         bool grilled;
         bool fried;
-        int cost;
+        double cost;
         std::string meal;
     public:
         Order();
         ~Order();
-        Order(int table, int waiter, std::vector<OrderItem> orderItems, bool grilled, bool fried, int cost);
+        Order(int table, int waiter, std::vector<OrderItem> orderItems, bool grilled, bool fried, double cost);
 
         bool contains(std::string item);
         void addItem(std::string item, double price, int quantity = 1);
-        void addItemCost(int costItem);
+        void addItemCost(double costItem);
         void removeItem(std::string item);
         void addDrink(std::string drink);
         std::string toStringItems();
@@ -32,7 +32,7 @@ class Order {
         void setOrderItems(std::vector<OrderItem> orderItems);
         void setGrilled(bool grilled);
         void setFried(bool fried);
-        void setCost(int cost);
+        void setCost(double cost);
         std::string getMeal();
         int getTable();
         int getWaiter();
